Natural Language Processing with Disaster Tweets
This is a competition where I have access to a dataset of 10,000 tweets that were hand classified and challenged to build a machine learning model that predicts which Tweets that are about real disasters and which one’s are not. Twitter has become an important communication channel in times of emergency. The ubiquitousness of smartphones enables people to announce an emergency they’re observing in real-time. Because of this, more agencies are interested in programatically monitoring Twitter. This competition has successfully submitted and predicted more than 3260 tweets with f1-score 76% accuracy.